// src/enums/navigation-menu.enum.ts
export enum NAVIGATION_MENU_TITLES {
  Order = 'ORDER',
  Cards = 'CARDS',
  Gift = 'GIFT',
  Stores = 'STORES'
}

export enum NAVIGATION_ORDER_MENU_TITLES {
  Menu = 'Menu',
  Favorite = 'Favorite',
  Previous = 'Previous'
}

export enum NavigationMenuRoutes {
  order = '/order',
  cards = '/cards',
  stores = '/stores',
  gift = '/gift'
}

export enum NavigationSubMenuRoutes {
  menu = '/order/menu',
  favorite = '/order/favorite',
  previous = '/order/previous'
}